added package filter. added ListIssueFilter.java added some apis in IssueApi.java
